TempPokeChange message now handled. This means ditto transformations now behave correctly, and so do moves that hide the sprite (e.g. fly, dig, dive). Additionally the battle now displays empty moves until a poke is sent out.

+++ src/com/pokebros/android/pokemononline/poke/UniqueID.java
@@ -13,9 +13,9 @@ public class UniqueID extends SerializeBytes {
 		subNum = msg.readByte();
 	}
 	
-	public UniqueID(short s, byte b) {
-		pokeNum = s;
-		subNum = b;
+	public UniqueID(int s, int b) {
+		pokeNum = (short) s;
+		subNum = (byte) b;
 	}
 	
 	public UniqueID() {
